+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Avatar reporting
 * AvatarMask reporting
 * Domain Reload Roslyn analyzer
+* New Area category for IterationTime issues
+* Code/Domain Reload view, showing any issues raised by the Domain Reload Roslyn analyzer
 
 ### Changed
 * Improved issue description and suggestion strings, to make them clearer and more consistent